+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed stack overflow when parsing malformed JSON. [#380](https://github.com/Microsoft/PSRule/issues/380)
 - Added rule documentation links and notes to help. [#382](https://github.com/Microsoft/PSRule/issues/382)
   - Added `-Full` switch to `Get-PSRuleHelp` to display links and notes sections.
+- Added aliases for `-OutputFormat` (`-o`) and `-Module` (`-m`) parameters. [#384](https://github.com/Microsoft/PSRule/issues/384)
 
 ## v0.13.0-B1912043 (pre-release)